sessions/next24/pulumi/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
pulumi-synced-folder>=0.0.0,<1.0.0
pulumi_google_native
requests>=2.0.0,<3.0.0
